requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
chardet==4.0.0
click==8.1.3
colorama==0.4.6
discord.py==1.7.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
networkx==2.6.3
nltk==3.6.2
numpy==1.21.2
regex==2022.10.31
scikit-learn==0.24.2
scipy==1.10.0
summa==1.2.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.5.0
yarl==1.8.2
